Abstracting the theme toggler into its own component.
